+++ gcpy/grid.py
@@ -1,10 +1,14 @@
+"""
+Module containing variables and functions that define and
+manipulate GEOS-Chem horizontal and vertical grids
+"""
 import xarray as xr
 import numpy as np
 import scipy.sparse
 from itertools import product
-from .util import get_shape_of_data
+from gcpy.util import get_shape_of_data, verify_variable_type
 from .grid_stretching_transforms import scs_transform
-from .constants import R_EARTH_m
+from gcpy.constants import R_EARTH_m
 
 
 def get_troposphere_mask(ds):
@@ -165,7 +169,7 @@ def call_make_grid(res, gridtype, in_extent=[-180, 180, -90, 90],
 
     Returns:
         [grid, grid_list]: list(dict, list(dict))
-            Returns the created grid. 
+            Returns the created grid.
             grid_list is a list of grids if gridtype is 'cs', else it is None
     """
 
@@ -286,6 +290,72 @@ def get_vert_grid(dataset, AP=[], BP=[]):
     else:
         new_grid = vert_grid(AP, BP)
         return new_grid.p_edge(), new_grid.p_mid(), np.size(AP)
+
+
+def get_lev_and_ilev(
+        n_lev=72,
+        AP_edge=None,
+        BP_edge=None
+):
+    """
+    Returns the eta values (defined as (A/P0) + B) at vertical
+    level edges and midpoints.  These are used to define the
+    "lev" and "ilev" netCDF coordinate values.
+
+    Keyword Args (optional):
+    ------------------------
+    n_lev : int
+        Number of levels in the grid.  Default = 72
+    AP_edge : list-like
+        Hybrid grid parameter A (hPa), with values placed on level
+        edges.  If not specified, values from the _GEOS_72L_AP array
+        in this module will be used.
+    AP_edge : list-like
+        Hybrid grid parameter B (unitless), with values placed on level
+        edges.  If not specified, values from the _GEOS_72L_BP array in
+        this module will be used.
+
+    Returns:
+    --------
+    lev : xr.DataArray
+        List of eta values at vertical grid midpoints
+    ilev : xr.DataArray
+        List of eta values at vertical grid edges
+    """
+    if n_lev is None:
+        n_lev = 72
+    if AP_edge is None and n_lev == 72:
+        AP_edge = _GEOS_72L_AP
+    if AP_edge is None and n_lev == 47:
+        AP_edge = _GEOS_47L_AP
+    if BP_edge is None and n_lev == 72:
+        BP_edge = _GEOS_72L_BP
+    if BP_edge is None and n_lev == 47:
+        BP_edge = _GEOS_47L_BP
+
+    # Edges
+    AP_edge = np.array(AP_edge)
+    BP_edge = np.array(BP_edge)
+    ilev = xr.DataArray(
+        np.array((AP_edge/1000.0) + BP_edge, np.float64),
+        attrs={
+            "long_name": "hybrid level at interfaces ((A/P0)+B)",
+            "units": "level"
+            }
+        )
+
+    # Midpoints
+    AP_mid = (AP_edge[0:n_lev:1] + AP_edge[1:n_lev+1:1]) * 0.5
+    BP_mid = (BP_edge[0:n_lev:1] + BP_edge[1:n_lev+1:1]) * 0.5
+    lev = xr.DataArray(
+        np.array((AP_mid/1000.0) + BP_mid, np.float64),
+        attrs={
+            "long_name": "hybrid level at midpoints ((A/P0)+B)",
+            "units": "level"
+            }
+        )
+
+    return lev, ilev
 
 
 def get_pressure_indices(pedge, pres_range):
@@ -1121,7 +1191,7 @@ class CSGrid(object):
             pp[:, i, j] = latlon_to_cartesian(
                 lambda_rad[i, j], theta_rad[i, j])
 
-        # Map the edges on the sphere back to the cube. 
+        # Map the edges on the sphere back to the cube.
         #Note that all intersections are at x = -rsq3
         # print("EDGES")
         for ij in range(1, c + 1):