+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added SLURM run script `gcpy/benchmark/benchmark_slurm.sh`
 - Added `gcpy/gcpy_plot_style` style sheet for title and label default settings
 - Added new cubed-sphere grid inquiry functions to `gcpy/cstools.py`
+- Added function `get_lev_and_ilev` to `gcpy/grid.py`
 
 ### Changed
 - Simplified the Github issues templates into two options: `new-feature-or-discussion.md` and `question-issue.md`